Concurrency, between processes, had a bug. Fixed.
Database class wasn't close the database pointer until it destroys.
So, if a process hold an instance of Manager (thus an instance of
Database), the database file remains lock.
Fixed by opening and closing the database file each time.
I also opened the databse file in read-only mode when we didn't
want to write anything into it, which makes performance a little better.